Accumulate total number of probing bytes sent as long.
It seems like it'd be possible to send more than 2 Gb of probing over the lifetime of a session.

+++ src/main/kotlin/org/jitsi/nlj/transform/node/outgoing/ProbingDataSender.kt
@@ -62,8 +62,8 @@ class ProbingDataSender(
     private var localVideoSsrc: Long? = null
 
     // Stats
-    private var numProbingBytesSentRtx: Int = 0
-    private var numProbingBytesSentDummyData: Int = 0
+    private var numProbingBytesSentRtx: Long = 0
+    private var numProbingBytesSentDummyData: Long = 0
 
     init {
         streamInformationStore.onRtpPayloadTypesChanged { currentRtpPayloadTypes ->
